+++ CHANGELOG.md
@@ -5,8 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.5] - Not Released
+## [1.4.*] - Unreleased
 ### Added
+### Changed
+### Fixed
+
+## [1.4.7]
+### Added
+### Changed
+- Modified id type for show API method: number | string
+### Fixed
+- transformResponse can handle null response
+
+## [1.4.5]
+### Added
+- Tests
 ### Changed
 - Removed unused dependencies
 ### Fixed